taxonomy-builder-app/demos/react-demo/backend/requirements.txt
fastapi
pydantic
uvicorn
typing
python-dotenv
openai==0.28

#also need to run chocolatoney, and choco install make for makefile testing